댓글삭제 기능 구현, 댓글이없습니다 버그

+++ src/client/js/comment.js
@@ -2,10 +2,9 @@ const formBtn = document.querySelector(".video__comment button");
 const form = document.querySelector(".video__comment");
 const textarea = form.querySelector("textarea");
 const videoPlay = document.querySelector(".video__play");
-const delBtn = document.querySelector(".video__comment__delete");
-
+const delBtn = document.querySelectorAll(".video__comment__delete");
+const commentList = document.querySelector(".video__comment__list");
 function fakeComment(text) {
-  const commentList = document.querySelector(".video__comment__list");
   const newCommentItem = document.createElement("li");
   const newCommentText = document.createElement("span");
   const newCommentDelBtn = document.createElement("a");
@@ -17,22 +16,48 @@ function fakeComment(text) {
   newCommentDelBtn.innerText = "❌";
   commentList.prepend(newCommentItem);
 }
+function checkText(text) {
+  if (!text) {
+    alert("댓글을 입력해주세요");
+  }
+}
+function removeComment() {
+  console.log("hi");
+  if (!commentList.children.length) {
+    const li = document.createElement("li");
+    li.clssName = "noComment";
+    li.innerText = "댓글이 없습니다";
+    commentList.appendChild(li);
+  } else {
+    const noComment = commentList.querySelector(".noComment");
+    if (noComment) {
+      commentList.remove(noComment);
+    }
+  }
+}
 
-delBtn.addEventListener("click", async () => {
-  const commentId = delBtn.dataset.commentId;
-  const userId = delBtn.dataset.userId;
-  try {
-    await fetch(`/user/${userId}/${commentId}/delete`, {
-      method: "DELETE",
-    });
-  } catch (err) {
-    console.log(err);
+commentList.addEventListener("click", async (e) => {
+  if (e.target.className === "video__comment__delete") {
+    const delBtn = e.target;
+    const commentId = delBtn.dataset.commentId;
+    const userId = delBtn.dataset.userId;
+    try {
+      await fetch(`/api/video/${userId}/${commentId}/delete`, {
+        method: "DELETE",
+      });
+      delBtn.parentElement.parentElement.removeChild(delBtn.parentElement);
+      removeComment();
+    } catch (err) {
+      console.log(err);
+    }
   }
 });
 
 formBtn.addEventListener("click", async (e) => {
+  removeComment();
   e.preventDefault();
   const text = textarea.value;
+  checkText(text);
   const videoId = videoPlay.dataset.id;
   if (text === "") {
     return;
